common: harmonize common mocks & conn mocks
