+++ tests/unit/conn-test/conn-test-new.c
@@ -12,6 +12,8 @@
  */
 
 #include "conn-test-common.h"
+#include "mocks-rdma_cm.h"
+#include "test-common.h"
 
 /*
  * new_test_peer_NULL - NULL peer is invalid
@@ -214,7 +216,6 @@ delete_test_destroy_cq_EAGAIN(void **unused)
 
 	/* configure mocks: */
 	expect_value(rdma_destroy_qp, id, MOCK_CM_ID);
-	expect_value(ibv_destroy_cq, cq, MOCK_IBV_CQ);
 	will_return(ibv_destroy_cq, EAGAIN);
 	expect_value(rdma_destroy_id, id, MOCK_CM_ID);
 	will_return(rdma_destroy_id, MOCK_OK);
@@ -247,7 +248,6 @@ delete_test_destroy_cq_EAGAIN_destroy_id_EAGAIN(void **unused)
 
 	/* configure mocks: */
 	expect_value(rdma_destroy_qp, id, MOCK_CM_ID);
-	expect_value(ibv_destroy_cq, cq, MOCK_IBV_CQ);
 	will_return(ibv_destroy_cq, EAGAIN); /* first error */
 	expect_value(rdma_destroy_id, id, MOCK_CM_ID);
 	will_return(rdma_destroy_id, EIO); /* second error */
@@ -284,7 +284,6 @@ delete_test_destroy_id_EAGAIN(void **unused)
 
 	/* configure mocks: */
 	expect_value(rdma_destroy_qp, id, MOCK_CM_ID);
-	expect_value(ibv_destroy_cq, cq, MOCK_IBV_CQ);
 	will_return(ibv_destroy_cq, MOCK_OK);
 	expect_value(rdma_destroy_id, id, MOCK_CM_ID);
 	will_return(rdma_destroy_id, EAGAIN);
