+++ tests/unit/conn-test/conn-test-next-event.c
@@ -11,6 +11,8 @@
  */
 
 #include "conn-test-common.h"
+#include "mocks-rdma_cm.h"
+#include "test-common.h"
 
 /*
  * next_event_test_conn_NULL - NULL conn is invalid
